flask-app/requirements.txt
Flask==2.3.2
faiss-cpu==1.7.4
numpy==1.24.3
Pillow==10.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.38.0
accelerate>=0.26.0
safetensors>=0.4.1
bitsandbytes>=0.41.0
pykospacing==0.5
soynlp==0.0.493